Fix NameError when trying to view menus from dashboard
When a user visited any of the Hyrax controllers from the dashboard
(e.g. works, collections, profile, etc.) the page gave a NameError:
`undefined local variable or method 'jobs_path'`

Controllers from the Hyrax engine don't appear to have access to
route helpers for controllers in the parent application (Tenejo),
so I replaced the helper with the literal path in the problem view.