Add initial branding pages
This commit provides the outline of a branding preview function by
providing an input form for key theming elements and a preview showing
how those settings would appear on a mock homepage.

The process works by injecting a dynamically generated css into the
header of all pages.  For the time being, I've scoped the css to only
affect the preview on the them page.

When we're ready to apply the styles site-wide, we can either change
the class names to match existing Hyrax classes, or create a cleaned-up
tenejo-specific version of the impacted layouts and views.

Style values are read from a `Theme` object.  The present code implements
a `#current_theme` class method that reads the values stored in the
`Theme` record with id `1`.  Saving that record persists the style settings
so they can be reloaded whenever the application is restarted.