Fix layout rendering error
The current code trys to find the path for the current logo using
`url_for` which seems to work in development, but not production.

This uses a more explict path call that seems to work in both
environments.

Fixes errors like
```
ActionView::Template::Error: undefined method `attachment_path' for #<#<Class:...
```