Update preflight to distinguish between single and multi-value fields
Active fedora raises errors like
>“ArgumentError: You attempted to set the property `date_modified’ of
to an enumerable value. However, this property is declared as singular.”

and

> “ArgumentError: You attempted to set the property `identifier’ of
to a scalar value. However, this property is declared as being multivalued.”

These updates allow each preflight class to inspect the corresponding
repository model and determine which fields should be scalar values and
which can be multi-valued.

This change allows multiple values to be imported from a CSV file by
packing them into a field using the default separator, currently: |~|