Update importer to create works
* Updates the PFWork class to inspect the Work class for
   attribute names rather than hard-coding them
* Adds code to CSV importer to create new Work objects
   but does not use the actor stack to create Works
* Updates fixtures to mactch field names generated from classes

Works created by the importer will do not exactly match works
created via the UI using the default actor stack.  We will need
to decide if we want to create works in the importer via a
trimmed-down actor stack or whether the additional differences
can be handled effectively in the importer code.