Add persistent preview values for themes
For each styling option in the Theme preview, add a a new value
prefixed with `preview_`.  These values are used to render the styling
for the preview and the non-prefixed values are used for rendering
the main site.

E.G.
`preview_primary_color` displays in the dashboard preview
`primary_color` displays on the main site

"Apply Changes" copies the settings from preview_ to the main values
"Reset to Defaults" resets preview values do defualt, you still need
to apply the changes if you want to reset the whole site to default
styling.