Show tenejo version info on production environments
When the application is running in a production environment only the
current release version or branch is displayed.

In non-production environments, more detailed git commit information
and delpoyment dates are displayed.

For lack of a better method to detect produciton environments, the
code assumes that hostnames without hyphens are production:
* tenejo = production (only display version)
* tenejo-dev = non-production (display detailed deploy info)
* localhost = non-production (display detailed commit info)