Define BasicMetadata properties in local codebase
Collections, Works, and FileSets all inlcude Hyrax::BasicMetadata
to define their propertis.  We were making local additons to add
a `primary_identifier` property to Collections and Works, and now
need to add it to FileSets too.

Rather than add the property in a third location, let's bring all
of the BasicMetadata definitions into the local code.  This puts
us in a better position to make future metadata refinements too.