Revert move of file setup for CSV test
On a local system, this setup runs and leaves the files in place,
so after one run of the test suite, the expected files are always
available for all tests.

When running in parallel on CI, the files don't persist and might
be running on a container that hasn't had the setup from other
CSV tests run yet.

So I'm adding back the explict file setup so CSVs validate as
expected.