Save rights_statement as URI in CSV importer
Hyrax expects the URI for the corresponding rights statement to be
stored in Fedora instead of the human-readable label.

We want the importer to accept the human-readable values, so this
change implements a lookup from the label provided to the URI
to be stored.

This change also extracts the parent work logic to its own
method to address a RuboCop method length error.