Refactor tests to avoid using "è"
We had used filenames with "è" and "œ" in them to test UTF-8 encoding
compatibility across the ingest chain.

It appears that some process in copying from local environments to
github to CircleCI handles the encoding for "œ" but not for "è".

This appears to be a CI pipeline issue and not an application or
infrastructure issue, so we're simplifying the test to avoid the
problem character.