Allow unique matches on primary_identifier
Including primary_identifier in Hyrax::BasicMetadataIndexer.stored_fields
caused it to ignore the symbol and multiple: false setting in
the property definition and cause it to be tokenized with a _tesim
suffix instead.

This change removes the stored_fields setting and updates all searches
to use the explicilty suffixed _ssi Solr fieldname instead.

This allows us to do exact and unique matches on `primary_identifier`
even when two or more primary_identifiers share common prefix segements.