Create files during CSV import
Bypass as much of the actor stack as possible.  The actor stack
introduces significant overhead attempting to infer changes that
might occur based on browser based add/edit/deletes that we don't
need to accomodate in the importer.

The current code  calls the IngestLocalFileJob which then spawns
CharacterizeJob and CreateDerivativesJob in order to ensure files
attached by the CSV importer have the same characteristics and
behaviors as files attached via the browser UI.