Set the default thumbnail and representative on import (#137)
When creating or updating a work, set the thumbnail to the first
(ordered) fileset attached to the work.  Do the same for
representative medaia to trigger generation of a IIIF manifest.

If the work already has a thumbnail, do not change it.

If the work already has a reprenentative media do not change it.