Assign noids to files without identifier in CSV
When a CSV does not specifiy an explicit identifier for a 'File'
row, assign a NOID (Nice Opaque IDentifier) as the identifier for
that file.

If the CSV provides and identifier, use the provided identifier.
NOTE: the system will still generate a NOID and use that for internal
references to the object.