Ensure imported works get added to admin sets and workflows
When a new work gets added to the repository via the importer, we
want to ensure that it gets set to the expected workflow state.

This requires that we add the work to an Admin Set and have at least
one workflow set up for that Admin Set.

This change adds all newly imported works to the Default Admin Set
with whichever workflow is configured for that admin set.

This change also removes redundant data from the nesting test CSV
to speed up the test suite.