Add a CSV export service (#190)
The service accepts the an Export job that includes one or
more identifiers for export.  The metadata for corresponding
collections and works is serialized to a CSV.  The CSV is
saved as an attachment to the Export job and can then be
downloaded at a future point.

NOTE: this change does not include the UI components which will
be required for users to access the service.