Fix autoloading bugs in development
**ISSUE**
The previous solution for re-loading STI Job models in development
was breaking multiple other parts of the application, frequently
triggering errors like
```
A copy of Hyrax::AbilityHelper has been removed from the module tree but is still active!
```
We have also experiences errors similar to `Export` is not a valid STI subclass of `Job`.

**SOLUTION**
Revert the code back to match an earlier state where the Job subclasses
lived in the models directory and were explicitly required by name.