Refactor exporter tests to decouple file IO from csv_generation
Having to save and read a file attachment everytime we want to test
the behavior of CSV generation adds a lot of overhead to tests.

This change separates generating the CSV output as an in memory string
from savinging it as an ActiveStorage attachment so we can test
each separately.

This change also adds an after-suite call to clean up all ActiveStorage
so we don't get leftover attachment files from tests.