Export files attached to exported works (#207)
Updates the importer to export links to the files attached to each
exported work.

Files are exported in the same ordering they are attached to the work.

In Hyrax, the file we want is the `:original_file` attached
to each FileSet.  The Hyrax downloads controller abstracts this so we
can just provide the `:id` of the desired FileSet.