Fix import status reporting for nested objects
**ISSUE**
The import status for nested collections and works was not updating
to "complete" despite the objects being successfully created and
the overall job being marked as completed.

**DIAGNOSIS**
If an object's parent had following siblings in the heirarchy, the search
method would overwrite the the succesfully found object with `nil`
when completing the search iteration on the previous ancestor level.

**FIX**
Only assign `found` on the initial match so that it is not overwritten
by nil values returned from searching subsequent children at the same level.