Add validation for "Resource Type" on Works and Collections
This commit:
* Adds resource type validation to collections and works
* Adds warnings created by active record validations to displayed warnings
* Updates the resource type validation to handle multi-valued entries

Up to now, we have only been validating `resource_type` on files,
which have the field defined, but don't show it or allow it to be
edited.

This change adds the same validation to Works and Collecitons. This
fixes a gap where admins could import invalid resource types via CSV.