Improve CodeClimate and Coveralls reporting
**CodeClimate**
We have GitHub and CircleCI both configured to run CodeClimate
independely and don't need to be running it twice.  This change
removes the related configuration form CircleCI.  Going forward,
we will continue to use GitHub actions to trigger CodeClimate.

**Coveralls**
We have not been getting coverage reporting since March 2022 due
to dependency incompatibilities with the `coveralls` gem.  The
gem is no longer being maintained.  The last update occured in 2019:
https://rubygems.org/gems/coveralls

The `coveralls_reborn` gem is being actively maintained by the same
authors and appears to be the direct replacement:
https://rubygems.org/gems/coveralls_reborn