Fix preflight exception when file row has no files (#271)
**ISSUE**
When importing a CSV without a "files" column, preflight was throwing
an exception when trying to upack nil files.

**SOLUTION**
We don't want to make "files" a required column, because it's totally
valid to import a CSV with only Collections.  This change detects when
`files` in a a row with object type "File" is nil or empty and creates
a placeholder error `PFFile` object to store the relevant row-level warnings.

**OTHER CHANGES**
This commit also splits the Preflight and PreFlightObj tests into two
files since the preflight spec file was getting pretty long.

see spec/lib/tenejo/pre_flight_obj_spec.rb:26-40 for the new tests.